# lib/api-zod/src/generated/types/attendanceRecord.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Student Support Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface AttendanceRecord {
  id: number;
  course: string;
  attended: number;
  total: number;
  percentage: number;
  status: string;
}