# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Student Support Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatInput {
  /** @minLength 2 */
  question: string;
  studentId?: string;
}