# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Student Support Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { Citation } from './citation';

export interface ChatMessage {
  id: number;
  role: string;
  content: string;
  createdAt: string;
  citations?: Citation[];
  toolsUsed?: string[];
}