# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Student Support Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { Citation } from './citation';

export interface ChatResponse {
  id: number;
  answer: string;
  citations: Citation[];
  toolsUsed: string[];
  memoryUsed: string[];
  intent: string;
  createdAt: string;
}