# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * AI Student Support Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentProfile } from './studentProfile';

export interface Dashboard {
  documents: number;
  questions: number;
  upcomingExams: number;
  attendance: number;
  notices: number;
  profile: StudentProfile;
}